requirements.txt
streamlit
torch
torchvision
numpy<2
opencv-python-headless
Pillow